+++ frontend/src/metabase/query_builder/components/view/sidebars/DatasetManagementSection/DatasetMetadataStrengthIndicator/DatasetMetadataStrengthIndicator.tsx
@@ -6,7 +6,7 @@ import Tooltip from "metabase/components/Tooltip";
 import { color } from "metabase/lib/colors";
 import { useHover } from "metabase/hooks/use-hover";
 import Question from "metabase-lib/lib/Question";
-import { getDatasetMetadataCompletenessPercentage } from "metabase-lib/lib/metadata/utils/metadata";
+import { getDatasetMetadataCompletenessPercentage } from "metabase-lib/lib/metadata/utils/models";
 
 import {
   Root,
